Some bugs are fixed in the CommunicateFragment

+++ app/src/main/java/com/kivous/attendanceroom/ui/fragments/CommunicateFragment.kt
@@ -5,9 +5,9 @@ import android.text.Editable
 import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
-import android.view.View.OnFocusChangeListener
 import android.view.ViewGroup
 import android.view.WindowManager
+import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
@@ -27,10 +27,8 @@ import com.kivous.attendanceroom.ui.viewmodels.AppViewModel
 import com.kivous.attendanceroom.utils.Response
 import com.kivous.attendanceroom.utils.glideCircle
 import com.kivous.attendanceroom.utils.gone
-import com.kivous.attendanceroom.utils.logD
 import com.kivous.attendanceroom.utils.setImageViewTint
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -60,8 +58,19 @@ class CommunicateFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.vBackArrow.setOnClickListener {
-            findNavController().navigateUp()
+            binding.recyclerView.stopScroll().let {
+                findNavController().navigateUp()
+            }
         }
+
+        val callback = object : OnBackPressedCallback(true) {
+            override fun handleOnBackPressed() {
+                binding.recyclerView.stopScroll().let {
+                    findNavController().navigateUp()
+                }
+            }
+        }
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, callback)
 
         sharedViewModel.sharedClassCode.observe(viewLifecycleOwner) { code ->
             viewModel.getClassRoomDetails(code)
@@ -83,17 +92,6 @@ class CommunicateFragment : Fragment() {
 
         glideCircle(auth.currentUser!!.photoUrl.toString(), binding.ivProfile)
         setUpRecyclerView()
-
-
-        binding.etMessage.onFocusChangeListener = OnFocusChangeListener { _, hasFocus ->
-            if (hasFocus) {
-                lifecycleScope.launch {
-                    delay(300)
-                    setUpRecyclerView()
-                }
-
-            }
-        }
 
         binding.etMessage.addTextChangedListener(object : TextWatcher {
             override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
@@ -119,6 +117,7 @@ class CommunicateFragment : Fragment() {
     }
 
     private fun chatAdapterViewController(holder: ChatAdapter.ViewHolder, model: Chat) {
+
         holder.apply {
             tvMessage.text = model.message.toString()
             tvName.text = model.name.toString()
@@ -127,23 +126,18 @@ class CommunicateFragment : Fragment() {
             glideCircle(model.photoUrl.toString(), ivProfile)
 
             if (model.email == auth.currentUser!!.email) {
-                lifecycleScope.launch {
-                    networkViewModel.isConnected.collectLatest { isConnected ->
-                        itemView.setOnLongClickListener {
-                            if (isConnected) {
-                                MaterialAlertDialogBuilder(requireContext()).setTitle("Want to delete?")
-                                    .setPositiveButton("Delete") { _, _ ->
-                                        sharedViewModel.sharedClassCode.observe(viewLifecycleOwner) {
-                                            viewModel.deleteChat(it, model.id.toString())
-                                        }
-                                    }.setNegativeButton("Cancel") { dialog, _ ->
-                                        dialog.dismiss()
-                                    }.setCancelable(true).show()
+                itemView.setOnLongClickListener {
+                    MaterialAlertDialogBuilder(requireContext()).setTitle("Want to delete?")
+                        .setPositiveButton("Delete") { _, _ ->
+                            sharedViewModel.sharedClassCode.observe(
+                                viewLifecycleOwner
+                            ) {
+                                viewModel.deleteChat(it, model.id.toString())
                             }
-                            true
-                        }
-
-                    }
+                        }.setNegativeButton("Cancel") { dialog, _ ->
+                            dialog.dismiss()
+                        }.setCancelable(true).show()
+                    true
                 }
             }
         }
@@ -163,28 +157,22 @@ class CommunicateFragment : Fragment() {
                                 auth, requireContext(), it1, ::chatAdapterViewController
                             )
                         }!!
-
                         binding.recyclerView.layoutManager = LinearLayoutManager(requireContext())
                         binding.recyclerView.adapter = adapter
 
-                        try {
-                            adapter.registerAdapterDataObserver(object :
-                                RecyclerView.AdapterDataObserver() {
-                                override fun onItemRangeInserted(
-                                    positionStart: Int, itemCount: Int
-                                ) {
-                                    super.onItemRangeInserted(positionStart, itemCount)
-                                    val lastVisiblePosition =
-                                        LinearLayoutManager(requireContext()).findLastCompletelyVisibleItemPosition()
-                                    if (lastVisiblePosition == -1 || positionStart >= adapter.itemCount - 1 && lastVisiblePosition == positionStart - 1) {
-                                        binding.recyclerView.scrollToPosition(positionStart)
-                                    }
+                        adapter.registerAdapterDataObserver(object :
+                            RecyclerView.AdapterDataObserver() {
+                            override fun onItemRangeInserted(
+                                positionStart: Int, itemCount: Int
+                            ) {
+                                super.onItemRangeInserted(positionStart, itemCount)
+                                val lastVisiblePosition =
+                                    LinearLayoutManager(requireContext()).findLastCompletelyVisibleItemPosition()
+                                if (lastVisiblePosition == -1 || positionStart >= adapter.itemCount - 1 && lastVisiblePosition == positionStart - 1) {
+                                    binding.recyclerView.scrollToPosition(positionStart)
                                 }
-                            })
-
-                        } catch (e: Exception) {
-                            logD(e.message.toString())
-                        }
+                            }
+                        })
 
                         adapter.startListening()
                     }
